+++ Channel.js
@@ -1,14 +1,14 @@
-import { c as _toConsumableArray, a as _objectSpread2, u as uuidv4$1, l as _defineProperty, b as _slicedToArray, e as LocalizationContext, f as _inherits, h as _createClass, i as _classCallCheck, j as _possibleConstructorReturn, k as _getPrototypeOf, m as _assertThisInitialized, w as withSendbirdContext } from './LocalizationContext-5c5b45a0.js';
+import { c as _toConsumableArray, a as _objectSpread2, u as uuidv4$1, b as _slicedToArray, l as _defineProperty, e as LocalizationContext, f as _inherits, h as _createClass, i as _classCallCheck, j as _possibleConstructorReturn, k as _getPrototypeOf, m as _assertThisInitialized, w as withSendbirdContext } from './LocalizationContext-5c5b45a0.js';
 import React, { useEffect, useCallback, useRef, useMemo, useState, useContext, Component, useReducer } from 'react';
 import PropTypes from 'prop-types';
-import { p as SEND_USER_MESSAGE, S as SEND_MESSAGE_START, q as SEND_FILE_MESSAGE, U as UPDATE_USER_MESSAGE, D as DELETE_MESSAGE, E as EmojiListItems, r as ImageRenderer, I as Icon, a as IconTypes, s as Loader, g as IconColors, b as Label, c as LabelTypography, d as LabelColors, C as ContextMenu, e as IconButton, h as MenuItems, i as MenuItem, n as UserProfileContext, A as Avatar, o as UserProfile, f as TextButton, M as Modal, L as LabelStringSet, P as PlaceHolder, t as PlaceHolderTypes, l as UserProfileProvider } from './index-a2b521ce.js';
+import { p as SEND_USER_MESSAGE, S as SEND_MESSAGE_START, q as SEND_FILE_MESSAGE, U as UPDATE_USER_MESSAGE, D as DELETE_MESSAGE, E as EmojiListItems, r as ImageRenderer, I as Icon, a as IconTypes, s as Loader, g as IconColors, b as Label, c as LabelTypography, d as LabelColors, C as ContextMenu, e as IconButton, h as MenuItems, i as MenuItem, n as UserProfileContext, A as Avatar, o as UserProfile, f as TextButton, M as Modal, P as PlaceHolder, t as PlaceHolderTypes, L as LabelStringSet, l as UserProfileProvider } from './index-c97add1b.js';
 import { a as getMessageCreatedAt$4, b as getSenderName$2, c as getSenderProfileUrl$2 } from './utils-53ba1773.js';
-import { C as ChannelAvatar } from './index-9dc7a088.js';
+import { C as ChannelAvatar } from './index-fa2a92d8.js';
 import format from 'date-fns/format';
 import { M as MessageStatusType } from './type-0296584d.js';
 import { t as truncate, g as getIsSentFromStatus$3 } from './utils-cfdeb084.js';
 import 'react-dom';
-import { i as isImage, a as isVideo, c as compareIds, u as unSupported, L as LinkLabel, D as DateSeparator, M as MessageInput, F as FileViewer } from './index-f7b03766.js';
+import { i as isImage, a as isVideo, c as compareIds, u as unSupported, L as LinkLabel, D as DateSeparator, M as MessageInput, F as FileViewer } from './index-f48d282b.js';
 import isSameDay from 'date-fns/isSameDay';
 import { c as copyToClipboard$1, g as getSenderProfileUrl$1, a as getSenderName$1, b as getMessageCreatedAt$5, d as getIsSentFromStatus$2 } from './utils-d7f59026.js';
 import formatDistanceToNowStrict from 'date-fns/formatDistanceToNowStrict';
@@ -66,8 +66,8 @@ var UNDEFINED = 'undefined';
 var SUCCEEDED = SendingMessageStatus.SUCCEEDED,
     FAILED = SendingMessageStatus.FAILED,
     PENDING = SendingMessageStatus.PENDING;
-var scrollIntoLast = function scrollIntoLast(selector) {
-  var intialTry = arguments.length > 1 && arguments[1] !== undefined ? arguments[1] : 0;
+var scrollIntoLast = function scrollIntoLast() {
+  var intialTry = arguments.length > 0 && arguments[0] !== undefined ? arguments[0] : 0;
   var MAX_TRIES = 10;
   var currentTry = intialTry;
 
@@ -76,13 +76,12 @@ var scrollIntoLast = function scrollIntoLast(selector) {
   }
 
   try {
-    var nodes = document.querySelectorAll(selector);
-    var last = nodes[nodes.length - 1];
-    last.scrollIntoView(false);
-    /** alignToTop: false */
+    var scrollDOM = document.querySelector('.sendbird-conversation__scroll-container'); // eslint-disable-next-line no-multi-assign
+
+    scrollDOM.scrollTop = scrollDOM.scrollHeight;
   } catch (error) {
     setTimeout(function () {
-      scrollIntoLast(selector, currentTry + 1);
+      scrollIntoLast(currentTry + 1);
     }, 500 * currentTry);
   }
 };
@@ -100,7 +99,7 @@ var pubSubHandler = function pubSubHandler(channelUrl, pubSub, dispatcher) {
   subscriber.set(SEND_USER_MESSAGE, pubSub.subscribe(SEND_USER_MESSAGE, function (msg) {
     var channel = msg.channel,
         message = msg.message;
-    scrollIntoLast('.sendbird-msg--scroll-ref');
+    scrollIntoLast();
 
     if (channel && channelUrl === channel.url) {
       dispatcher({
@@ -123,7 +122,7 @@ var pubSubHandler = function pubSubHandler(channelUrl, pubSub, dispatcher) {
   subscriber.set(SEND_FILE_MESSAGE, pubSub.subscribe(SEND_FILE_MESSAGE, function (msg) {
     var channel = msg.channel,
         message = msg.message;
-    scrollIntoLast('.sendbird-msg--scroll-ref');
+    scrollIntoLast();
 
     if (channel && channelUrl === channel.url) {
       dispatcher({
@@ -296,6 +295,21 @@ var passUnsuccessfullMessages = function passUnsuccessfullMessages(allMessages, 
   }
 
   return [].concat(_toConsumableArray(allMessages), [newMessage]);
+};
+var pxToNumber = function pxToNumber(px) {
+  if (typeof px === 'number') {
+    return px;
+  }
+
+  if (typeof px === 'string') {
+    var parsed = Number.parseFloat(px);
+
+    if (!Number.isNaN(parsed)) {
+      return parsed;
+    }
+  }
+
+  return null;
 };
 
 var messagesInitialState = {
@@ -743,6 +757,9 @@ function useInitialMessagesFetch(_ref, _ref2) {
         currentGroupChannel: currentGroupChannel,
         userFilledMessageListQuery: userFilledMessageListQuery
       });
+      messagesDispatcher({
+        type: GET_PREV_MESSAGES_START
+      });
       currentGroupChannel.getMessagesByTimestamp(new Date().getTime(), messageListParams).then(function (messages) {
         var hasMore = messages && messages.length > 0;
         var lastMessageTimeStamp = hasMore ? messages[0].createdAt : null;
@@ -769,7 +786,7 @@ function useInitialMessagesFetch(_ref, _ref2) {
       }).finally(function () {
         currentGroupChannel.markAsRead();
         setTimeout(function () {
-          return scrollIntoLast('.sendbird-msg--scroll-ref');
+          return scrollIntoLast();
         });
       });
     }
@@ -828,7 +845,7 @@ function useHandleReconnect(_ref, _ref2) {
               }
             });
             setTimeout(function () {
-              return scrollIntoLast('.sendbird-msg--scroll-ref');
+              return scrollIntoLast();
             });
           }).catch(function (error) {
             logger.error('Channel: Fetching messages failed', error);
@@ -1144,7 +1161,7 @@ function useSendMessageCallback(_ref, _ref2) {
       channel: currentGroupChannel
     });
     setTimeout(function () {
-      return scrollIntoLast('.sendbird-msg--scroll-ref');
+      return scrollIntoLast();
     });
   }, [currentGroupChannel, onBeforeSendUserMessage]);
   return [messageInputRef, sendMessage];
@@ -1152,69 +1169,161 @@ function useSendMessageCallback(_ref, _ref2) {
 
 function useSendFileMessageCallback(_ref, _ref2) {
   var currentGroupChannel = _ref.currentGroupChannel,
-      onBeforeSendFileMessage = _ref.onBeforeSendFileMessage;
+      onBeforeSendFileMessage = _ref.onBeforeSendFileMessage,
+      _ref$imageCompression = _ref.imageCompression,
+      imageCompression = _ref$imageCompression === void 0 ? {} : _ref$imageCompression;
   var sdk = _ref2.sdk,
       logger = _ref2.logger,
       pubSub = _ref2.pubSub,
       messagesDispatcher = _ref2.messagesDispatcher;
   var sendMessage = useCallback(function (file) {
+    var compressionRate = imageCompression.compressionRate,
+        resizingWidth = imageCompression.resizingWidth,
+        resizingHeight = imageCompression.resizingHeight;
+    var createCustomParams = onBeforeSendFileMessage && typeof onBeforeSendFileMessage === 'function';
+    var compressibleFileType = file.type === 'image/jpg' || file.type === 'image/png' || file.type === 'image/jpeg';
+    var compressibleRatio = compressionRate > 0 && compressionRate < 1; // pxToNumber returns null if values are invalid
+
+    var compressibleDiamensions = pxToNumber(resizingWidth) || pxToNumber(resizingHeight);
+    var canCompressImage = compressibleFileType && (compressibleRatio || compressibleDiamensions);
+
     var createParamsDefault = function createParamsDefault(file_) {
       var params = new sdk.FileMessageParams();
       params.file = file_;
       return params;
     };
 
-    var createCustomPrams = onBeforeSendFileMessage && typeof onBeforeSendFileMessage === 'function';
+    if (canCompressImage) {
+      // Using image compression
+      try {
+        var image = document.createElement('img');
+        image.src = URL.createObjectURL(file);
 
-    if (createCustomPrams) {
-      logger.info('Channel: creating params using onBeforeSendFileMessage', onBeforeSendFileMessage);
-    }
+        image.onload = function () {
+          URL.revokeObjectURL(image.src);
+          var canvas = document.createElement('canvas');
+          var imageWdith = image.naturalWidth || image.width;
+          var imageHeight = image.naturalHeight || image.height;
+          var targetWidth = pxToNumber(resizingWidth) || imageWdith;
+          var targetHeight = pxToNumber(resizingHeight) || imageHeight; // In canvas.toBlob(callback, mimeType, qualityArgument)
+          // qualityArgument doesnt work
+          // so in case compressibleDiamensions are not present, we use ratio
 
-    var params = onBeforeSendFileMessage ? onBeforeSendFileMessage(file) : createParamsDefault(file);
-    logger.info('Channel: Uploading file message start!', params);
-    var pendingMsg = currentGroupChannel.sendFileMessage(params, function (response, err) {
-      var swapParams = sdk.getErrorFirstCallback();
-      var message = response;
-      var error = err;
+          if (file.type === 'image/png' && !compressibleDiamensions) {
+            targetWidth *= compressionRate;
+            targetHeight *= compressionRate;
+          }
 
-      if (swapParams) {
-        message = err;
-        error = response;
+          canvas.width = targetWidth;
+          canvas.height = targetHeight;
+          var context = canvas.getContext('2d');
+          context.drawImage(image, 0, 0, targetWidth, targetHeight);
+          context.canvas.toBlob(function (newImageBlob) {
+            var compressedFile = new File([newImageBlob], file.name, {
+              type: file.type
+            });
+
+            if (createCustomParams) {
+              logger.info('Channel: Creating params using onBeforeSendFileMessage', onBeforeSendFileMessage);
+            }
+
+            var params = createCustomParams ? onBeforeSendFileMessage(compressedFile) : createParamsDefault(compressedFile);
+            logger.info('Channel: Uploading file message start!', params);
+            var pendingMessage = currentGroupChannel.sendFileMessage(params, function (response, err) {
+              var swapParams = sdk.getErrorFirstCallback();
+
+              var _ref3 = swapParams ? [err, response] : [response, err],
+                  _ref4 = _slicedToArray(_ref3, 2),
+                  message = _ref4[0],
+                  error = _ref4[1];
+
+              if (error) {
+                // sending params instead of pending message
+                // to make sure that we can resend the message once it fails
+                logger.error('Channel: Sending file message failed!', message);
+                message.localUrl = URL.createObjectURL(compressedFile);
+                message.file = compressedFile;
+                messagesDispatcher({
+                  type: SEND_MESSAGEGE_FAILURE,
+                  payload: message
+                });
+                return;
+              }
+
+              logger.info('Channel: Sending file message success!', message);
+              messagesDispatcher({
+                type: SEND_MESSAGEGE_SUCESS,
+                payload: message
+              });
+            });
+            pubSub.publish(SEND_MESSAGE_START, {
+              /* pubSub is used instead of messagesDispatcher
+                to avoid redundantly calling `messageActionTypes.SEND_MESSAGEGE_START` */
+              message: _objectSpread2({}, pendingMessage, {
+                url: URL.createObjectURL(compressedFile),
+                // pending thumbnail message seems to be failed
+                requestState: 'pending'
+              }),
+              channel: currentGroupChannel
+            });
+            setTimeout(function () {
+              return scrollIntoLast();
+            }, 1000);
+          }, file.type, compressionRate);
+        };
+      } catch (error) {
+        logger.error('Channel: Sending file message failed!');
+      }
+    } else {
+      // Not using image compression
+      if (createCustomParams) {
+        logger.info('Channel: creating params using onBeforeSendFileMessage', onBeforeSendFileMessage);
       }
 
-      if (error) {
-        // sending params instead of pending message
-        // to make sure that we can resend the message once it fails
-        logger.error('Channel: Sending file message failed!', message);
-        message.localUrl = URL.createObjectURL(file);
-        message.file = file;
+      var params = onBeforeSendFileMessage ? onBeforeSendFileMessage(file) : createParamsDefault(file);
+      logger.info('Channel: Uploading file message start!', params);
+      var pendingMsg = currentGroupChannel.sendFileMessage(params, function (response, err) {
+        var swapParams = sdk.getErrorFirstCallback();
+
+        var _ref5 = swapParams ? [err, response] : [response, err],
+            _ref6 = _slicedToArray(_ref5, 2),
+            message = _ref6[0],
+            error = _ref6[1];
+
+        if (error) {
+          // sending params instead of pending message
+          // to make sure that we can resend the message once it fails
+          logger.error('Channel: Sending file message failed!', message);
+          message.localUrl = URL.createObjectURL(file);
+          message.file = file;
+          messagesDispatcher({
+            type: SEND_MESSAGEGE_FAILURE,
+            payload: message
+          });
+          return;
+        }
+
+        logger.info('Channel: Sending message success!', message);
         messagesDispatcher({
-          type: SEND_MESSAGEGE_FAILURE,
+          type: SEND_MESSAGEGE_SUCESS,
           payload: message
         });
-        return;
-      }
-
-      logger.info('Channel: Sending message success!', message);
-      messagesDispatcher({
-        type: SEND_MESSAGEGE_SUCESS,
-        payload: message
       });
-    });
-    pubSub.publish(SEND_MESSAGE_START, {
-      /* pubSub is used instead of messagesDispatcher
-        to avoid redundantly calling `messageActionTypes.SEND_MESSAGEGE_START` */
-      message: _objectSpread2({}, pendingMsg, {
-        url: URL.createObjectURL(file),
-        // pending thumbnail message seems to be failed
-        requestState: 'pending'
-      }),
-      channel: currentGroupChannel
-    });
-    setTimeout(function () {
-      return scrollIntoLast('.sendbird-msg--scroll-ref');
-    }, 1000);
-  }, [currentGroupChannel, onBeforeSendFileMessage]);
+      pubSub.publish(SEND_MESSAGE_START, {
+        /* pubSub is used instead of messagesDispatcher
+          to avoid redundantly calling `messageActionTypes.SEND_MESSAGEGE_START` */
+        message: _objectSpread2({}, pendingMsg, {
+          url: URL.createObjectURL(file),
+          // pending thumbnail message seems to be failed
+          requestState: 'pending'
+        }),
+        channel: currentGroupChannel
+      });
+      setTimeout(function () {
+        return scrollIntoLast();
+      }, 1000);
+    }
+  }, [currentGroupChannel, onBeforeSendFileMessage, imageCompression]);
   return [sendMessage];
 }
 
@@ -1898,6 +2007,11 @@ function OutgoingUserMessage(_ref) {
       moreActive = _useState4[0],
       setMoreActive = _useState4[1];
 
+  var _useState5 = useState(false),
+      _useState6 = _slicedToArray(_useState5, 2),
+      menuDisplaying = _useState6[0],
+      setMenuDisplaying = _useState6[1];
+
   var handleMoreIconClick = function handleMoreIconClick() {
     setMoreActive(true);
   };
@@ -1934,6 +2048,7 @@ function OutgoingUserMessage(_ref) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -1945,12 +2060,20 @@ function OutgoingUserMessage(_ref) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: parentRefMenus // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: parentContainRef // for toggling more options(menus & reactions)
-        ,
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: parentRefMenus,
+        parentContainRef: parentContainRef,
         closeDropdown: closeDropdown,
         openLeft: true
       }, isMessageSent && React.createElement(MenuItem, {
@@ -1994,6 +2117,7 @@ function OutgoingUserMessage(_ref) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2005,7 +2129,12 @@ function OutgoingUserMessage(_ref) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         message: message,
         parentRef: parentRefReactions,
@@ -2016,14 +2145,11 @@ function OutgoingUserMessage(_ref) {
         }
       });
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement("div", {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement("div", {
     className: "sendbird-user-message__status"
   }, React.createElement(MessageStatus, {
     message: message,
-    status: status // onDelete={() => { showRemove(true); }}
-    // onResend={() => resendMessage(message)}
-    // duplicated and should replace to more
-
+    status: status
   }))), React.createElement("div", {
     className: "sendbird-user-message__text-balloon"
   }, React.createElement("div", {
@@ -2068,15 +2194,20 @@ function IncomingUserMessage(_ref2) {
       disableUserProfile = _React$useContext.disableUserProfile,
       renderUserProfile = _React$useContext.renderUserProfile;
 
-  var _useState5 = useState(false),
-      _useState6 = _slicedToArray(_useState5, 2),
-      mousehover = _useState6[0],
-      setMousehover = _useState6[1];
-
   var _useState7 = useState(false),
       _useState8 = _slicedToArray(_useState7, 2),
-      moreActive = _useState8[0],
-      setMoreActive = _useState8[1];
+      mousehover = _useState8[0],
+      setMousehover = _useState8[1];
+
+  var _useState9 = useState(false),
+      _useState10 = _slicedToArray(_useState9, 2),
+      moreActive = _useState10[0],
+      setMoreActive = _useState10[1];
+
+  var _useState11 = useState(false),
+      _useState12 = _slicedToArray(_useState11, 2),
+      menuDisplaying = _useState12[0],
+      setMenuDisplaying = _useState12[1];
 
   var showReactionAddButton = useReaction && emojiAllMap && emojiAllMap.size > 0;
   var showEmojiReactions = useReaction && message.reactions && message.reactions.length > 0;
@@ -2120,11 +2251,14 @@ function IncomingUserMessage(_ref2) {
       });
     },
     menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: avatarRef // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: avatarRef // for toggling more options(menus & reactions)
-        ,
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: avatarRef,
+        parentContainRef: avatarRef,
         closeDropdown: closeDropdown,
         style: {
           paddingTop: 0,
@@ -2164,7 +2298,10 @@ function IncomingUserMessage(_ref2) {
     className: "sendbird-user-message__right-padding"
   }, React.createElement("div", {
     className: "sendbird-user-message__more",
-    ref: parentContainRef
+    ref: parentContainRef,
+    style: {
+      top: chainTop ? '6px' : '22px'
+    }
   }, showReactionAddButton && React.createElement(ContextMenu, {
     menuTrigger: function menuTrigger(toggleDropdown) {
       return React.createElement(IconButton, {
@@ -2174,6 +2311,7 @@ function IncomingUserMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2185,7 +2323,12 @@ function IncomingUserMessage(_ref2) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         parentRef: parentRefReactions,
         parentContainRef: parentContainRef,
@@ -2205,6 +2348,7 @@ function IncomingUserMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2216,7 +2360,12 @@ function IncomingUserMessage(_ref2) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MenuItems, {
         parentRef: parentRefMenus,
         parentContainRef: parentContainRef,
@@ -2229,7 +2378,7 @@ function IncomingUserMessage(_ref2) {
         }
       }, "Copy"));
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement(Label, {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement(Label, {
     className: "sendbird-user-message__sent-at",
     type: LabelTypography.CAPTION_3,
     color: LabelColors.ONBACKGROUND_2
@@ -2392,7 +2541,9 @@ function OutgoingThumbnailMessage(_ref2) {
       chainBottom = _ref2.chainBottom;
   var type = message.type,
       url = message.url,
-      localUrl = message.localUrl;
+      localUrl = message.localUrl,
+      thumbnails = message.thumbnails;
+  var thumbnailUrl = thumbnails && thumbnails.length > 0 && thumbnails[0].url || null;
 
   var _useContext = useContext(LocalizationContext),
       stringSet = _useContext.stringSet;
@@ -2411,6 +2562,11 @@ function OutgoingThumbnailMessage(_ref2) {
       _useState4 = _slicedToArray(_useState3, 2),
       moreActive = _useState4[0],
       setMoreActive = _useState4[1];
+
+  var _useState5 = useState(false),
+      _useState6 = _slicedToArray(_useState5, 2),
+      menuDisplaying = _useState6[0],
+      setMenuDisplaying = _useState6[1];
 
   var showReactionAddButton = useReaction && emojiAllMap && emojiAllMap.size > 0;
   var MemoizedEmojiListItems = memoizedEmojiListItems;
@@ -2451,6 +2607,7 @@ function OutgoingThumbnailMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2462,8 +2619,18 @@ function OutgoingThumbnailMessage(_ref2) {
         color: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
         parentRef: menuRef,
         parentContainRef: parentContainRef,
         closeDropdown: closeDropdown,
@@ -2493,6 +2660,7 @@ function OutgoingThumbnailMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2504,7 +2672,12 @@ function OutgoingThumbnailMessage(_ref2) {
         color: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         message: message,
         parentRef: reactionAddRef,
@@ -2515,7 +2688,7 @@ function OutgoingThumbnailMessage(_ref2) {
         }
       });
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement(MessageStatus, {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement(MessageStatus, {
     message: message,
     status: status,
     className: "".concat(OUTGOING_THUMBNAIL_MESSAGE, "-left-padding__status")
@@ -2533,7 +2706,15 @@ function OutgoingThumbnailMessage(_ref2) {
     } : function () {},
     tabIndex: 0,
     role: "button"
-  }, isVideo(type) && React.createElement(React.Fragment, null, React.createElement("video", {
+  }, isVideo(type) && React.createElement(React.Fragment, null, thumbnailUrl ? React.createElement(ImageRenderer, {
+    className: "".concat(OUTGOING_THUMBNAIL_MESSAGE, "-body__video"),
+    url: thumbnailUrl,
+    alt: "video/thumbnail",
+    width: "404px",
+    height: "280px"
+  }) :
+  /* eslint-disable-next-line jsx-a11y/media-has-caption */
+  React.createElement("video", {
     className: "".concat(OUTGOING_THUMBNAIL_MESSAGE, "-body__video")
   }, React.createElement("source", {
     src: url || localUrl,
@@ -2543,16 +2724,12 @@ function OutgoingThumbnailMessage(_ref2) {
     width: "56px",
     height: "56px",
     type: IconTypes.PLAY
-  })), isImage(type) && React.createElement("div", {
+  })), isImage(type) && React.createElement(ImageRenderer, {
     className: "".concat(OUTGOING_THUMBNAIL_MESSAGE, "-body__img"),
-    style: {
-      backgroundImage: "url(".concat(url || localUrl, ")"),
-      height: '280px',
-      width: '404px',
-      backgroundSize: 'cover',
-      backgroundRepeat: 'no-repeat',
-      backgroundPosition: 'center'
-    }
+    url: thumbnailUrl || url || localUrl,
+    alt: "image/thumbnail",
+    width: "404px",
+    height: "280px"
   }), unSupported(type) && React.createElement("div", {
     className: "".concat(OUTGOING_THUMBNAIL_MESSAGE, "-body__other")
   }, stringSet.UNKNOWN__UNKNOWN_MESSAGE_TYPE), React.createElement("div", {
@@ -2582,7 +2759,9 @@ function IncomingThumbnailMessage(_ref3) {
       chainBottom = _ref3.chainBottom;
   var type = message.type,
       url = message.url,
-      localUrl = message.localUrl;
+      localUrl = message.localUrl,
+      thumbnails = message.thumbnails;
+  var thumbnailUrl = thumbnails && thumbnails.length > 0 && thumbnails[0].url || null;
 
   var _React$useContext = React.useContext(UserProfileContext),
       disableUserProfile = _React$useContext.disableUserProfile,
@@ -2596,15 +2775,20 @@ function IncomingThumbnailMessage(_ref3) {
   var reactionAddRef = useRef(null);
   var avatarRef = useRef(null);
 
-  var _useState5 = useState(false),
-      _useState6 = _slicedToArray(_useState5, 2),
-      mousehover = _useState6[0],
-      setMousehover = _useState6[1];
-
   var _useState7 = useState(false),
       _useState8 = _slicedToArray(_useState7, 2),
-      moreActive = _useState8[0],
-      setMoreActive = _useState8[1];
+      mousehover = _useState8[0],
+      setMousehover = _useState8[1];
+
+  var _useState9 = useState(false),
+      _useState10 = _slicedToArray(_useState9, 2),
+      moreActive = _useState10[0],
+      setMoreActive = _useState10[1];
+
+  var _useState11 = useState(false),
+      _useState12 = _slicedToArray(_useState11, 2),
+      menuDisplaying = _useState12[0],
+      setMenuDisplaying = _useState12[1];
 
   var showReactionAddButton = useReaction && emojiAllMap && emojiAllMap.size > 0;
   var MemoizedEmojiListItems = memoizedEmojiListItems;
@@ -2642,6 +2826,7 @@ function IncomingThumbnailMessage(_ref3) {
   }, !chainBottom && React.createElement(ContextMenu, {
     menuTrigger: function menuTrigger(toggleDropdown) {
       return React.createElement(Avatar, {
+        ref: avatarRef,
         onClick: function onClick() {
           if (!disableUserProfile) {
             toggleDropdown();
@@ -2654,11 +2839,14 @@ function IncomingThumbnailMessage(_ref3) {
       });
     },
     menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: avatarRef // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: avatarRef // for toggling more options(menus & reactions)
-        ,
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: avatarRef,
+        parentContainRef: avatarRef,
         closeDropdown: closeDropdown,
         style: {
           paddingTop: 0,
@@ -2682,7 +2870,15 @@ function IncomingThumbnailMessage(_ref3) {
     onKeyDown: isMessageSent ? function () {
       return onClick(true);
     } : function () {}
-  }, isVideo(type) && React.createElement(React.Fragment, null, React.createElement("video", {
+  }, isVideo(type) && React.createElement(React.Fragment, null, thumbnailUrl ? React.createElement(ImageRenderer, {
+    className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__video"),
+    src: thumbnailUrl,
+    alt: "video/thumbnail",
+    width: "404px",
+    height: "280px"
+  }) :
+  /* eslint-disable-next-line jsx-a11y/media-has-caption */
+  React.createElement("video", {
     className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__video")
   }, React.createElement("source", {
     src: url || localUrl,
@@ -2692,16 +2888,12 @@ function IncomingThumbnailMessage(_ref3) {
     width: "56px",
     height: "56px",
     type: IconTypes.PLAY
-  })), isImage(type) && React.createElement("div", {
+  })), isImage(type) && React.createElement(ImageRenderer, {
     className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__img"),
-    style: {
-      backgroundImage: "url(".concat(url || localUrl, ")"),
-      height: '280px',
-      width: '404px',
-      backgroundSize: 'cover',
-      backgroundRepeat: 'no-repeat',
-      backgroundPosition: 'center'
-    }
+    url: thumbnailUrl || url || localUrl,
+    alt: "image/thumbnail",
+    width: "404px",
+    height: "280px"
   }), unSupported(type) && React.createElement("div", {
     className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__other")
   }, stringSet.UNKNOWN__UNKNOWN_MESSAGE_TYPE), React.createElement("div", {
@@ -2716,7 +2908,7 @@ function IncomingThumbnailMessage(_ref3) {
     memoizedEmojiListItems: memoizedEmojiListItems
   }))), React.createElement("div", {
     className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__right-padding")
-  }, !chainBottom && !(mousehover || moreActive) && React.createElement(Label, {
+  }, !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement(Label, {
     className: "".concat(INCOMING_THUMBNAIL_MESSAGE, "__sent-at"),
     type: LabelTypography.CAPTION_3,
     color: LabelColors.ONBACKGROUND_2
@@ -2732,6 +2924,7 @@ function IncomingThumbnailMessage(_ref3) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2743,7 +2936,12 @@ function IncomingThumbnailMessage(_ref3) {
         color: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         message: message,
         parentRef: reactionAddRef,
@@ -2893,6 +3091,11 @@ function OutgoingFileMessage(_ref) {
       moreActive = _useState4[0],
       setMoreActive = _useState4[1];
 
+  var _useState5 = useState(false),
+      _useState6 = _slicedToArray(_useState5, 2),
+      menuDisplaying = _useState6[0],
+      setMenuDisplaying = _useState6[1];
+
   var showReactionAddButton = useReaction && emojiAllMap && emojiAllMap.size > 0;
   var MemoizedEmojiListItems = memoizedEmojiListItems;
   var isMessageSent = getIsSentFromStatus$3(status);
@@ -2933,6 +3136,7 @@ function OutgoingFileMessage(_ref) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2943,8 +3147,18 @@ function OutgoingFileMessage(_ref) {
         type: IconTypes.MORE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
         parentRef: menuRef,
         parentContainRef: parentContainRef,
         closeDropdown: closeDropdown,
@@ -2974,6 +3188,7 @@ function OutgoingFileMessage(_ref) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -2985,7 +3200,12 @@ function OutgoingFileMessage(_ref) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         message: message,
         parentRef: reactionAddButtonRef,
@@ -2996,7 +3216,7 @@ function OutgoingFileMessage(_ref) {
         }
       });
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement("div", {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement("div", {
     className: "sendbird-file-message__status"
   }, React.createElement(MessageStatus, {
     message: message,
@@ -3051,15 +3271,20 @@ function IncomingFileMessage(_ref2) {
   var avatarRef = useRef(null);
   var reactionAddButtonRef = useRef(null);
 
-  var _useState5 = useState(false),
-      _useState6 = _slicedToArray(_useState5, 2),
-      mousehover = _useState6[0],
-      setMousehover = _useState6[1];
-
   var _useState7 = useState(false),
       _useState8 = _slicedToArray(_useState7, 2),
-      moreActive = _useState8[0],
-      setMoreActive = _useState8[1];
+      mousehover = _useState8[0],
+      setMousehover = _useState8[1];
+
+  var _useState9 = useState(false),
+      _useState10 = _slicedToArray(_useState9, 2),
+      moreActive = _useState10[0],
+      setMoreActive = _useState10[1];
+
+  var _useState11 = useState(false),
+      _useState12 = _slicedToArray(_useState11, 2),
+      menuDisplaying = _useState12[0],
+      setMenuDisplaying = _useState12[1];
 
   var showReactionAddButton = useReaction && emojiAllMap && emojiAllMap.size > 0;
   var MemoizedEmojiListItems = memoizedEmojiListItems;
@@ -3103,11 +3328,14 @@ function IncomingFileMessage(_ref2) {
       });
     },
     menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: avatarRef // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: avatarRef // for toggling more options(menus & reactions)
-        ,
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: avatarRef,
+        parentContainRef: avatarRef,
         closeDropdown: closeDropdown,
         style: {
           paddingTop: 0,
@@ -3162,6 +3390,7 @@ function IncomingFileMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -3173,7 +3402,12 @@ function IncomingFileMessage(_ref2) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
       return React.createElement(MemoizedEmojiListItems, {
         message: message,
         parentRef: reactionAddButtonRef,
@@ -3184,7 +3418,7 @@ function IncomingFileMessage(_ref2) {
         }
       });
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement(Label, {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement(Label, {
     className: "sendbird-file-message__sent-at",
     type: LabelTypography.CAPTION_3,
     color: LabelColors.ONBACKGROUND_2
@@ -3409,6 +3643,11 @@ function OutgoingUnknownMessage(_ref2) {
       moreActive = _useState4[0],
       setMoreActive = _useState4[1];
 
+  var _useState5 = useState(false),
+      _useState6 = _slicedToArray(_useState5, 2),
+      menuDisplaying = _useState6[0],
+      setMenuDisplaying = _useState6[1];
+
   var _useContext = useContext(LocalizationContext),
       stringSet = _useContext.stringSet;
 
@@ -3448,6 +3687,7 @@ function OutgoingUnknownMessage(_ref2) {
         onClick: function onClick() {
           toggleDropdown();
           handleMoreIconClick();
+          setMenuDisplaying(true);
         },
         onBlur: function onBlur() {
           handleMoreIconBlur();
@@ -3459,12 +3699,20 @@ function OutgoingUnknownMessage(_ref2) {
         fillColor: IconColors.CONTENT_INVERSE
       }));
     },
-    menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: menuRef // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: parentContainRef // for toggling more options(menus & reactions)
-        ,
+    menuItems: function menuItems(close) {
+      var closeDropdown = function closeDropdown() {
+        close();
+        setMenuDisplaying(false);
+      };
+
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: menuRef,
+        parentContainRef: parentContainRef,
         closeDropdown: closeDropdown,
         openLeft: true
       }, React.createElement(MenuItem, {
@@ -3474,7 +3722,7 @@ function OutgoingUnknownMessage(_ref2) {
         }
       }, "Delete"));
     }
-  })), !chainBottom && !(mousehover || moreActive) && React.createElement("div", {
+  })), !chainBottom && !(mousehover || moreActive || menuDisplaying) && React.createElement("div", {
     className: "".concat(className, "__message-status")
   }, React.createElement(MessageStatus, {
     message: message,
@@ -3536,11 +3784,14 @@ function IncomingUnknownMessage(_ref3) {
       });
     },
     menuItems: function menuItems(closeDropdown) {
-      return React.createElement(MenuItems, {
-        parentRef: avatarRef // for catching location(x, y) of MenuItems
-        ,
-        parentContainRef: avatarRef // for toggling more options(menus & reactions)
-        ,
+      return React.createElement(MenuItems
+      /**
+       * parentRef: For catching location(x, y) of MenuItems
+       * parentContainRef: For toggling more options(menus & reactions)
+       */
+      , {
+        parentRef: avatarRef,
+        parentContainRef: avatarRef,
         closeDropdown: closeDropdown,
         style: {
           paddingTop: 0,
@@ -4552,7 +4803,7 @@ MessageHoc.propTypes = {
     ogMetaData: PropTypes.shape({})
   }),
   renderCustomMessage: PropTypes.func,
-  currentGroupChannel: PropTypes.shape,
+  currentGroupChannel: PropTypes.shape({}),
   hasSeperator: PropTypes.bool,
   disabled: PropTypes.bool,
   editDisabled: PropTypes.bool,
@@ -4672,6 +4923,21 @@ function (_Component) {
           useMessageGrouping = _this$props2.useMessageGrouping,
           currentGroupChannel = _this$props2.currentGroupChannel,
           memoizedEmojiListItems = _this$props2.memoizedEmojiListItems;
+
+      if (!initialized) {
+        return React.createElement(PlaceHolder, {
+          className: "sendbird-conversation__not-initialized",
+          type: PlaceHolderTypes.WRONG
+        });
+      }
+
+      if (allMessages.length < 1) {
+        return React.createElement(PlaceHolder, {
+          className: "sendbird-conversation__no-messages",
+          type: PlaceHolderTypes.NO_MESSAGES
+        });
+      }
+
       return React.createElement("div", {
         className: "sendbird-conversation__messages"
       }, React.createElement("div", {
@@ -4682,7 +4948,7 @@ function (_Component) {
         className: "sendbird-conversation__padding"
       }), React.createElement("div", {
         className: "sendbird-conversation__messages-padding"
-      }, initialized && allMessages.map(function (m, idx) {
+      }, allMessages.map(function (m, idx) {
         var previousMessage = allMessages[idx - 1];
         var nextMessage = allMessages[idx + 1];
 
@@ -5148,6 +5414,7 @@ var ConversationPanel = function ConversationPanel(props) {
       pubSub = _props$config.pubSub,
       isOnline = _props$config.isOnline,
       theme = _props$config.theme,
+      imageCompression = _props$config.imageCompression,
       reconnect = props.dispatchers.reconnect,
       _props$queries = props.queries,
       queries = _props$queries === void 0 ? {} : _props$queries,
@@ -5195,10 +5462,11 @@ var ConversationPanel = function ConversationPanel(props) {
       emojiContainer = messagesStore.emojiContainer,
       readStatus = messagesStore.readStatus;
   var isFrozen = currentGroupChannel.isFrozen,
-      isBroadcast = currentGroupChannel.isBroadcast;
+      isBroadcast = currentGroupChannel.isBroadcast,
+      isSuper = currentGroupChannel.isSuper;
   var _sdk$appInfo = sdk.appInfo,
       appInfo = _sdk$appInfo === void 0 ? {} : _sdk$appInfo;
-  var usingReaction = appInfo.isUsingReaction && !isBroadcast || useReaction || false;
+  var usingReaction = appInfo.isUsingReaction && !isBroadcast && !isSuper && useReaction;
   var userDefinedDisableUserProfile = disableUserProfile || config.disableUserProfile;
   var userDefinedRenderProfile = renderUserProfile || config.renderUserProfile;
   var emojiAllMap = useMemo(function () {
@@ -5324,7 +5592,8 @@ var ConversationPanel = function ConversationPanel(props) {
 
   var _useSendFileMessageCa = useSendFileMessageCallback({
     currentGroupChannel: currentGroupChannel,
-    onBeforeSendFileMessage: onBeforeSendFileMessage
+    onBeforeSendFileMessage: onBeforeSendFileMessage,
+    imageCompression: imageCompression
   }, {
     sdk: sdk,
     logger: logger,
@@ -5388,7 +5657,7 @@ var ConversationPanel = function ConversationPanel(props) {
   }), isFrozen && React.createElement(FrozenNotification, null), unreadCount > 0 && React.createElement(Notification, {
     count: unreadCount,
     onClick: function onClick() {
-      scrollIntoLast('.sendbird-msg--scroll-ref'); // there is no scroll
+      scrollIntoLast(); // there is no scroll
 
       if (scrollRef.current.scrollTop === 0) {
         currentGroupChannel.markAsRead();
@@ -5485,6 +5754,11 @@ ConversationPanel.propTypes = {
     pubSub: PropTypes.shape({
       subscribe: PropTypes.func,
       publish: PropTypes.func
+    }),
+    imageCompression: PropTypes.shape({
+      compressionRate: PropTypes.number,
+      resizingWidth: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
+      resizingHeight: PropTypes.oneOfType([PropTypes.number, PropTypes.string])
     })
   }).isRequired,
   queries: PropTypes.shape({